pyproject.toml
[project]
name = "pastpuzzle-rss"
version = "0.1.0"
description = "Daily PastPuzzle RSS feed generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12",
  "click>=8.1",
  "httpx>=0.27",
  "lxml>=5.2",
  "playwright>=1.57.0",
  "python-dotenv>=1.0",
]

[dependency-groups]
test = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
